requirements.txt
flask
flask-sqlalchemy
pyodbc
flask-migrate
python-dotenv
pywin32
fpdf
requests